Remove copied refactoring classes that exist in jdt.core.manipulation.
