+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/corext/refactoring/structure/ReferenceFinderUtil.java
@@ -35,7 +35,7 @@ import org.eclipse.jdt.core.dom.Type;
 import org.eclipse.jdt.core.search.SearchEngine;
 import org.eclipse.jdt.core.search.SearchMatch;
 import org.eclipse.jdt.internal.corext.refactoring.CollectingSearchRequestor;
-import org.eclipse.jdt.ls.core.internal.corext.util.SearchUtils;
+import org.eclipse.jdt.internal.corext.util.SearchUtils;
 
 public class ReferenceFinderUtil {
